requirements.txt
asgiref==3.3.1
autopep8==1.5.4
boto3==1.16.63
botocore==1.19.63
certifi==2020.12.5
chardet==4.0.0
Django==3.1.6
django-appconf==1.0.4
django-compressor==2.4
django-debug-toolbar==3.2
django-el-pagination==3.3.0
django-libsass==0.8
django-mathfilters==1.0.0
django-silk==4.1.0
django-storages==1.11.1
djangorestframework==3.12.2
freeze==3.0
gprof2dot==2019.11.30
idna==3.1
Jinja2==2.11.3
jmespath==0.10.0
libsass==0.20.1
MarkupSafe==1.1.1
Pillow==8.1.0
psycopg2==2.8.6
pycodestyle==2.6.0
Pygments==2.7.4
python-dateutil==2.8.1
python-gettext==4.0
python-magic==0.4.18
pytz==2021.1
rcssmin==1.0.6
requests==2.25.1
rjsmin==1.1.0
s3cmd==2.1.0
s3transfer==0.3.4
selenium==3.141.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.3
